@@ -15,30 +15,30 @@ describe ChannelGroupsController do
       get :new, {}
       expect(response).to redirect_to new_user_session_path
 
-      get :show, {:id => channel_group.to_param}
+      get :show, { id: channel_group.to_param }
       expect(response).to redirect_to new_user_session_path
 
-      get :messages_report, {:id => channel_group.to_param}
+      get :messages_report, { id: channel_group.to_param }
       expect(response).to redirect_to new_user_session_path
 
-      get :edit, {:id => channel_group.to_param}
+      get :edit, { id: channel_group.to_param }
       expect(response).to redirect_to new_user_session_path
 
       expect {
-            post :create, {:channel_group => valid_attributes}
+            post :create, { channel_group: valid_attributes }
           }.to_not change(ChannelGroup, :count)
 
       expect_any_instance_of(ChannelGroup).not_to receive(:update_attributes)
-      put :update, {:id => channel_group.to_param, :channel_group => { "name" => "MyString" }}
+      put :update, { id: channel_group.to_param, channel_group: { name: "MyString" }}
 
       expect {
           delete :destroy, {:id => channel_group.to_param}
       }.to_not change(ChannelGroup, :count)
 
-      channel = create(:channel,user:user)
+      channel = create(:channel, user:user)
       channel_group.channels << channel
       expect {
-        post :remove_channel, {channel_group_id:channel_group.to_param, id:channel.to_param }
+        post :remove_channel, { channel_group_id: channel_group.id, id: channel.id }
       }.to_not change(channel_group.channels, :count)
 
     end
@@ -50,27 +50,27 @@ describe ChannelGroupsController do
       another_user = create(:user)
       sign_in another_user
 
-      get :show, {:id => channel_group.to_param}
+      get :show, { id: channel_group.id }
       expect(response).to redirect_to root_url
 
-      get :messages_report, {:id => channel_group.to_param}
+      get :messages_report, { id: channel_group.id }
       expect(response).to redirect_to root_url
 
 
-      get :edit, {:id => channel_group.to_param}
+      get :edit, { id: channel_group.id }
       expect(response).to redirect_to root_url
 
       expect_any_instance_of(ChannelGroup).not_to receive(:update_attributes)
-      put :update, {:id => channel_group.to_param, :channel => { "name" => "MyString" }}
+      put :update, {id: channel_group.id, channel: { name: "MyString" }}
 
       expect {
-          delete :destroy, {:id => channel_group.to_param}
+          delete :destroy, { id: channel_group.id }
       }.to_not change(ChannelGroup, :count)
 
-      channel = create(:channel,user:user)
+      channel = create(:channel, user:user)
       channel_group.channels << channel
       expect {
-        post :remove_channel, {channel_group_id:channel_group.to_param, id:channel.to_param }
+        post :remove_channel, { channel_group_id: channel_group.id, id: channel.id }
       }.to_not change(channel_group.channels, :count)
     end
   end
@@ -83,16 +83,16 @@ describe ChannelGroupsController do
         channel_group = user.channel_groups.create! valid_attributes
         channel_group = user.channel_groups.find(channel_group.id)
         channels = (0..2).map {create(:channel,user:user)}
-        channels = (0..2).map {|i| Channel.find(channels[i])}
+        channels = (0..2).map {|i| Channel.find(channels[i].id)}
         channel_group.channels << channels
-        get :show, {user_id:user.to_param, :id => channel_group.to_param}
+        get :show, { user_id:user.id, id: channel_group.id }
         expect(assigns(:channel_group)).to eq(channel_group)
         expect(assigns(:channels)).to match(channels)
       end
     end
     describe "GET new" do
       it "assigns a new channel as @channel" do
-        get :new, {user_id:user.to_param}
+        get :new, { user_id: user.id }
         expect(assigns(:channel_group)).to be_a_new(ChannelGroup)
       end
     end
@@ -100,7 +100,7 @@ describe ChannelGroupsController do
       it "assigns the requested channel_group as @channel_group" do
         channel_group = user.channel_groups.create! valid_attributes
         channel_group = ChannelGroup.find(channel_group.id)
-        get :edit, {user_id:user.to_param,:id => channel_group.to_param}
+        get :edit, { user_id: user.id, id: channel_group.id }
         expect(assigns(:channel_group)).to eq(channel_group)
       end
     end
@@ -108,18 +108,18 @@ describe ChannelGroupsController do
       describe "with valid params" do
         it "creates a new ChannelGroup" do
           expect {
-            post :create, {:channel_group => valid_attributes}
+            post :create, { :channel_group => valid_attributes }
           }.to change(ChannelGroup, :count).by(1)
         end
 
         it "assigns a newly created channel group as @channel_group" do
-          post :create, {:channel_group => valid_attributes}
+          post :create, { :channel_group => valid_attributes }
           expect(assigns(:channel_group)).to be_a(ChannelGroup)
           expect(assigns(:channel_group)).to be_persisted
         end
 
         it "redirects to the created channel_group" do
-          post :create, {user_id:user.to_param,:channel_group => valid_attributes}
+          post :create, { user_id:user.id, :channel_group => valid_attributes }
           expect(response).to redirect_to(ChannelGroup.last)
         end
       end
@@ -127,13 +127,13 @@ describe ChannelGroupsController do
       describe "with invalid params" do
         it "assigns a newly created but unsaved channel as @channel" do
           allow_any_instance_of(ChannelGroup).to receive(:save).and_return(false)
-          post :create, {user_id:user.to_param, :channel_group => { "name" => "invalid value" }}
+          post :create, {user_id: user.id, :channel_group => { "name" => "invalid value" }}
           expect(assigns(:channel_group)).to be_a_new(ChannelGroup)
         end
 
         it "re-renders the 'new' template" do
           allow_any_instance_of(ChannelGroup).to receive(:save).and_return(false)
-          post :create, {user_id:user.to_param, :channel_group => { "name" => "invalid value" }}
+          post :create, {user_id: user.id, :channel_group => { "name" => "invalid value" }}
           expect(response).to render_template("new")
         end
       end
@@ -143,20 +143,20 @@ describe ChannelGroupsController do
         it "updates the requested channel_group" do
           channel_group = user.channel_groups.create! valid_attributes
           expect_any_instance_of(ChannelGroup).to receive(:update_attributes).with({ "name" => "MyString" })
-          put :update, {:id => channel_group.to_param, :channel_group => { "name" => "MyString" }}
+          put :update, {:id => channel_group.id, :channel_group => { "name" => "MyString" }}
         end
 
         it "assigns the requested channel_group as @channel_group" do
           channel_group = user.channel_groups.create! valid_attributes
           channel_group = ChannelGroup.find(channel_group.id)
-          put :update, {:id => channel_group.to_param, :channel_group => valid_attributes}
+          put :update, {:id => channel_group.id, :channel_group => valid_attributes}
           expect(assigns(:channel_group)).to eq(channel_group)
         end
 
         it "redirects to the channel" do
           channel_group = user.channel_groups.create! valid_attributes
           channel_group = ChannelGroup.find(channel_group.id)
-          put :update, {:id => channel_group.to_param, :channel_group => valid_attributes}
+          put :update, {:id => channel_group.id, :channel_group => valid_attributes}
           expect(response).to redirect_to(channel_group)
         end
       end
@@ -166,14 +166,14 @@ describe ChannelGroupsController do
           channel_group = user.channel_groups.create! valid_attributes
           channel_group = ChannelGroup.find(channel_group.id)
           allow_any_instance_of(ChannelGroup).to receive(:save).and_return(false)
-          put :update, {:id => channel_group.to_param, :channel_group => { "name" => "invalid value" }}
+          put :update, {:id => channel_group.id, :channel_group => { "name" => "invalid value" }}
           expect(assigns(:channel_group)).to eq(channel_group)
         end
 
         it "re-renders the 'edit' template" do
           channel_group = user.channel_groups.create! valid_attributes
           allow_any_instance_of(ChannelGroup).to receive(:save).and_return(false)
-          put :update, {:id => channel_group.to_param, :channel_group => { "name" => "invalid value" }}
+          put :update, {:id => channel_group.id, :channel_group => { "name" => "invalid value" }}
           expect(response).to render_template("edit")
         end
       end
@@ -182,13 +182,13 @@ describe ChannelGroupsController do
       it "destroys the requested channel_group" do
         channel_group = user.channel_groups.create! valid_attributes
         expect {
-          delete :destroy, {:id => channel_group.to_param}
+          delete :destroy, {:id => channel_group.id}
         }.to change(ChannelGroup, :count).by(-1)
       end
 
       it "redirects to the user path" do
         channel_group = user.channel_groups.create! valid_attributes
-        delete :destroy, {:id => channel_group.to_param}
+        delete :destroy, { :id => channel_group.id }
         expect(response).to redirect_to(user_path(user))
       end
     end
