@@ -1,11 +1,13 @@
 require 'spec_helper'
 
 describe  SubscriberActivitiesController do
+
   it "does not recognize the new,create and destory actions" do
-    expect {get :new, {}}.to raise_error(ActionController::RoutingError)
-    expect {post :create, {}}.to raise_error(ActionController::RoutingError)
-    expect {delete :destroy, {}}.to raise_error(ActionController::RoutingError)
+    expect {get :new, {}}.to raise_error(ActionController::UrlGenerationError)
+    expect {post :create, {}}.to raise_error(ActionController::UrlGenerationError)
+    expect {delete :destroy, {}}.to raise_error(ActionController::UrlGenerationError)
   end
+
   let(:user)                      { create(:user)                     }
   let(:channel)                   { create(:channel,user:user)        }
   let(:channel_group)             { create(:channel_group, user:user) }
@@ -22,20 +24,20 @@ describe  SubscriberActivitiesController do
 
   describe "guest user" do
     it "is redirected to signup form always and not allowed to alter db" do
-      get :index, {subscriber_id:subscriber}
+      get :index, { subscriber_id: subscriber.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :show, {subscriber_id:subscriber,:id => subscriber_response.to_param}
+      get :show, { subscriber_id: subscriber.id, id: subscriber_response.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :edit, {subscriber_id:subscriber,:id => subscriber_response.to_param}
+      get :edit, { subscriber_id: subscriber.id, id: subscriber_response.id }
       expect(response).to redirect_to new_user_session_path
 
       expect_any_instance_of(SubscriberResponse).not_to receive(:update_attributes)
-      put :update, {subscriber_id:subscriber,:id => subscriber_response.to_param, :caption => "Some Caption"}
+      put :update, { subscriber_id: subscriber.id, id: subscriber_response.id, caption: 'Some Caption' }
 
       expect_any_instance_of(SubscriberResponse).not_to receive(:process)
-      post :reprocess, {subscriber_id:subscriber,:id => subscriber_response.to_param, :caption => "Some Caption"}
+      post :reprocess, { subscriber_id: subscriber.id, id: subscriber_response.id, caption: 'Some Caption' }
 
     end
   end
@@ -45,20 +47,20 @@ describe  SubscriberActivitiesController do
       another_user = create(:user)
       sign_in another_user
 
-      get :index, {subscriber_id:subscriber}
+      get :index, { subscriber_id: subscriber.id }
       expect(response).to redirect_to root_url
 
-      get :show, {subscriber_id:subscriber,:id => subscriber_response.to_param}
+      get :show, {subscriber_id: subscriber.id, id: subscriber_response.id }
       expect(response).to redirect_to root_url
 
-      get :edit, {subscriber_id:subscriber,:id => subscriber_response.to_param}
+      get :edit, { subscriber_id: subscriber.id, id: subscriber_response.id }
       expect(response).to redirect_to root_url
 
       expect_any_instance_of(SubscriberActivity).not_to receive(:update_attributes)
-      put :update, {subscriber_id:subscriber,:id => subscriber_response.to_param, :caption => "Some Caption"}
+      put :update, { subscriber_id: subscriber.id, id: subscriber_response.id, caption: 'Some Caption' }
 
       expect_any_instance_of(SubscriberActivity).not_to receive(:process)
-      post :reprocess, {subscriber_id:subscriber,:id => subscriber_response.to_param, :caption => "Some Caption"}
+      post :reprocess, {subscriber_id:subscriber.id, id: subscriber_response.id, caption: 'Some Caption' }
 
     end
   end
@@ -69,23 +71,23 @@ describe  SubscriberActivitiesController do
     end
     describe "GET index" do
       it "when called with subscriber id,  lists all his activities" do
-        get :index, {subscriber_id:subscriber}
-        expect(assigns(:subscriber_activities)).to eq([SubscriberActivity.find(subscriber_response)])
+        get :index, { subscriber_id: subscriber.id }
+        expect(assigns(:subscriber_activities)).to eq([SubscriberActivity.find(subscriber_response.id)])
       end
 
       it "when called for a message, lists all its subscriber activities" do
-        get :index, {message_id:message,channel_id:channel}
-        expect(assigns(:subscriber_activities)).to eq([SubscriberActivity.find(subscriber_response)])
+        get :index, { message_id: message.id, channel_id: channel.id }
+        expect(assigns(:subscriber_activities)).to eq([SubscriberActivity.find(subscriber_response.id)])
       end
 
       it "when called for a channel, lists all its subscriber activities" do
-        get :index, {channel_id:channel}
-        expect(assigns(:subscriber_activities)).to match_array([SubscriberActivity.find(subscriber_response),
-                    SubscriberActivity.find(other_subscriber_response)])
+        get :index, { channel_id: channel.id }
+        expect(assigns(:subscriber_activities)).to match_array([SubscriberActivity.find(subscriber_response.id),
+                    SubscriberActivity.find(other_subscriber_response.id)])
       end
 
       it "when called for a channel_group, lists all its subscriber activities" do
-        get :index, {channel_group_id:channel_group}
+        get :index, { channel_group_id: channel_group.id }
         expect(assigns(:subscriber_activities)).to match_array([SubscriberActivity.find(channel_group_response.id)])
       end
 
@@ -93,39 +95,39 @@ describe  SubscriberActivitiesController do
 
     describe "GET show" do
       it "assigns the requested message's subscriber activity as @subscriber_activity" do
-        get :show, {message_id:message,channel_id:channel,:id => subscriber_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response))
+        get :show, { message_id: message.id, channel_id: channel.id, id: subscriber_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response.id))
       end
       it "assigns the requested subscriber's subscriber activity as @subscriber_activity" do
-        get :show, {subscriber_id:subscriber,:id => subscriber_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response))
+        get :show, { subscriber_id: subscriber.id, id: subscriber_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response.id))
       end
       it "assigns the requested channel's subscriber activity as @subscriber_activity" do
-        get :show, {channel_id:channel,:id => subscriber_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response))
+        get :show, { channel_id:channel.id, id: subscriber_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response.id))
       end
       it "assigns the requested channel_group's subscriber activity as @subscriber_activity" do
-        get :show, {channel_group_id:channel_group,:id => channel_group_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(channel_group_response))
+        get :show, { channel_group_id:channel_group.id, id: channel_group_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(channel_group_response.id))
       end
     end
 
     describe "GET edit" do
       it "assigns the requested message's subscriber activity as @subscriber_activity" do
-        get :edit, {message_id:message,channel_id:channel,:id => subscriber_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response))
+        get :edit, { message_id: message.id, channel_id:channel.id, id: subscriber_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response.id))
       end
       it "assigns the requested subscriber's subscriber activity as @subscriber_activity" do
-        get :edit, {subscriber_id:subscriber,:id => subscriber_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response))
+        get :edit, { subscriber_id: subscriber.id, id: subscriber_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response.id))
       end
       it "assigns the requested channel's subscriber activity as @subscriber_activity" do
-        get :edit, {channel_id:channel,:id => subscriber_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response))
+        get :edit, { channel_id: channel.id,  id: subscriber_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(subscriber_response.id))
       end
       it "assigns the requested channel_group's subscriber activity as @subscriber_activity" do
-        get :edit, {channel_group_id:channel_group,:id => channel_group_response.to_param}
-        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(channel_group_response))
+        get :edit, { channel_group_id: channel_group.id, id: channel_group_response.to_param }
+        expect(assigns(:subscriber_activity)).to eq(SubscriberActivity.find(channel_group_response.id))
       end
     end
 
@@ -133,23 +135,23 @@ describe  SubscriberActivitiesController do
       describe "with valid params" do
         it "updates the requested message's subscriber activity" do
           expect_any_instance_of(SubscriberResponse).to receive(:update_attributes).with({ "caption" => "Sample Caption" })
-          put :update, {message_id:message,channel_id:channel,:id => subscriber_response.to_param, :subscriber_activity=>{"caption" => "Sample Caption"} }
+          put :update, { message_id: message.id, channel_id:channel.id, id: subscriber_response.to_param, subscriber_activity: {caption:  'Sample Caption'} }
         end
         it "updates the requested subscriber's subscriber activity" do
           expect_any_instance_of(SubscriberResponse).to receive(:update_attributes).with({ "caption" => "Sample Caption" })
-          put :update, {subscriber_id:subscriber,:id => subscriber_response.to_param, :subscriber_activity=>{"caption" => "Sample Caption"} }
+          put :update, { subscriber_id: subscriber.id, id: subscriber_response.to_param, subscriber_activity: {caption: 'Sample Caption'} }
         end
         it "updates the requested channel's subscriber activity" do
           expect_any_instance_of(SubscriberResponse).to receive(:update_attributes).with({ "caption" => "Sample Caption" })
-          put :update, {channel_id:channel,:id => subscriber_response.to_param, :subscriber_activity=>{"caption" => "Sample Caption"} }
+          put :update, { channel_id: channel.id, id: subscriber_response.to_param, subscriber_activity: {caption: 'Sample Caption'} }
         end
         it "updates the requested channel group's subscriber activity" do
           expect_any_instance_of(SubscriberResponse).to receive(:update_attributes).with({ "caption" => "Sample Caption" })
-          put :update, {channel_group_id:channel_group,:id => channel_group_response.to_param, :subscriber_activity=>{"caption" => "Sample Caption"} }
+          put :update, { channel_group_id: channel_group.id, id: channel_group_response.to_param, subscriber_activity: { caption: 'Sample Caption'} }
         end
 
         it "redirects to the subscriber activity" do
-          put :update, {message_id:message,channel_id:channel,:id => subscriber_response.to_param, :subscriber_activity=>{"caption" => "Sample Caption"}}
+          put :update, { message_id:message.id, channel_id:channel.id, id: subscriber_response.to_param, subscriber_activity: {caption: 'Sample Caption'}}
           expect(response).to redirect_to :action => :show, message_id:message,channel_id:channel,id:subscriber_response.to_param
         end
       end
@@ -157,7 +159,7 @@ describe  SubscriberActivitiesController do
       describe "with invalid params" do
         it "re-renders the 'edit' template" do
           allow_any_instance_of(SubscriberResponse).to receive(:save).and_return(false)
-          put :update, {message_id:message,channel_id:channel,:id => subscriber_response.to_param, :subscriber_activity=>{"caption" => "Sample Caption"}}
+          put :update, { message_id: message.id, channel_id:channel.id, id: subscriber_response.to_param, subscriber_activity: {caption: 'Sample Caption'}}
           expect(response).to render_template("edit")
         end
      end
@@ -165,7 +167,7 @@ describe  SubscriberActivitiesController do
     describe "POST reprocess" do
       it "reprocesses the requested response" do
         expect_any_instance_of(SubscriberResponse).to receive(:process){}
-        post :reprocess, {message_id:message,channel_id:channel,:id => subscriber_response.to_param, :subscriber_activity=>{"caption" => "Sample Caption"} }
+        post :reprocess, { message_id: message.id, channel_id: channel.id, id: subscriber_response.to_param, subscriber_activity: {caption: 'Sample Caption'} }
       end
     end
   end
