@@ -18,13 +18,13 @@ describe ChannelsController do
       get :new, {}
       expect(response).to redirect_to new_user_session_path
 
-      get :show, {:id => channel.to_param}
+      get :show, { id: channel.id}
       expect(response).to redirect_to new_user_session_path
 
-      get :messages_report, {:id => channel.to_param}
+      get :messages_report, { id: channel.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :edit, {:id => channel.to_param}
+      get :edit, {:id => channel.id }
       expect(response).to redirect_to new_user_session_path
 
       expect {
@@ -32,24 +32,24 @@ describe ChannelsController do
           }.to_not change(Channel, :count)
 
       expect_any_instance_of(Channel).not_to receive(:update_attributes)
-      put :update, {:id => channel.to_param, :channel => { "name" => "MyString" }}
+      put :update, {:id => channel.id, :channel => { "name" => "MyString" }}
 
       expect {
-          delete :destroy, {:id => channel.to_param}
+          delete :destroy, {:id => channel.id}
       }.to_not change(Channel, :count)
 
-      get :list_subscribers, {:id => channel.to_param}
+      get :list_subscribers, {:id => channel.id}
       expect(response).to redirect_to new_user_session_path
 
       subscriber = create(:subscriber,user:user)
       expect {
-        post :add_subscriber, {channel_id:channel.to_param, id:subscriber.to_param }
+        post :add_subscriber, {channel_id:channel.id, id:subscriber.id }
       }.to_not change(channel.subscribers, :count)
 
       subscriber = create(:subscriber,user:user)
       channel.subscribers << subscriber
       expect {
-        post :remove_subscriber, {channel_id:channel.to_param, id:subscriber.to_param }
+        post :remove_subscriber, {channel_id:channel.id, id:subscriber.id }
       }.to_not change(channel.subscribers, :count)
 
     end
@@ -61,34 +61,34 @@ describe ChannelsController do
       another_user = create(:user)
       sign_in another_user
 
-      get :show, {:id => channel.to_param}
+      get :show, {:id => channel.id}
       expect(response).to redirect_to root_url
 
-      get :messages_report, {:id => channel.to_param}
+      get :messages_report, {:id => channel.id}
       expect(response).to redirect_to root_url
 
-      get :edit, {:id => channel.to_param}
+      get :edit, {:id => channel.id}
       expect(response).to redirect_to root_url
 
       expect_any_instance_of(Channel).not_to receive(:update_attributes)
-      put :update, {:id => channel.to_param, :channel => { "name" => "MyString" }}
+      put :update, {:id => channel.id, :channel => { "name" => "MyString" }}
 
       expect {
-          delete :destroy, {:id => channel.to_param}
+          delete :destroy, {:id => channel.id}
       }.to_not change(Channel, :count)
 
-      get :list_subscribers, {:id => channel.to_param}
+      get :list_subscribers, {:id => channel.id}
       expect(response).to redirect_to root_url
 
       subscriber = create(:subscriber,user:user)
       expect {
-        post :add_subscriber, {channel_id:channel.to_param, id:subscriber.to_param }
+        post :add_subscriber, {channel_id:channel.id, id:subscriber.id }
       }.to_not change(channel.subscribers, :count)
 
       subscriber = create(:subscriber,user:user)
       channel.subscribers << subscriber
       expect {
-        post :remove_subscriber, {channel_id:channel.to_param, id:subscriber.to_param }
+        post :remove_subscriber, {channel_id:channel.id, id:subscriber.id }
       }.to_not change(channel.subscribers, :count)
 
     end
@@ -121,14 +121,14 @@ describe ChannelsController do
         channel = user.channels.find(channel.id)
         subscribers = (0..2).map {create(:subscriber,user:user)}
         channel.subscribers << subscribers
-        get :show, {user_id:user.to_param, :id => channel.to_param}
+        get :show, {user_id:user.id, :id => channel.id}
         expect(assigns(:channel)).to eq(channel)
         expect(assigns(:subscribers)).to match(subscribers)
       end
     end
     describe "GET new" do
       it "assigns a new channel as @channel" do
-        get :new, {user_id:user.to_param}
+        get :new, {user_id:user.id}
         expect(assigns(:channel)).to be_a_new(Channel)
       end
     end
@@ -136,7 +136,7 @@ describe ChannelsController do
       it "assigns the requested channel as @channel" do
         channel = user.channels.create! valid_attributes
         channel = Channel.find(channel.id)
-        get :edit, {user_id:user.to_param,:id => channel.to_param}
+        get :edit, {user_id:user.id,:id => channel.id}
         expect(assigns(:channel)).to eq(channel)
       end
     end
@@ -155,7 +155,7 @@ describe ChannelsController do
         end
 
         it "redirects to the created channel" do
-          post :create, {user_id:user.to_param,:channel => valid_attributes}
+          post :create, {user_id:user.id,:channel => valid_attributes}
           expect(response).to redirect_to(Channel.last)
         end
       end
@@ -164,14 +164,14 @@ describe ChannelsController do
         it "assigns a newly created but unsaved channel as @channel" do
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Channel).to receive(:save).and_return(false)
-          post :create, {user_id:user.to_param, :channel => { "name" => "invalid value" }}
+          post :create, {user_id:user.id, :channel => { "name" => "invalid value" }}
           expect(assigns(:channel)).to be_a_new(Channel)
         end
 
         it "re-renders the 'new' template" do
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Channel).to receive(:save).and_return(false)
-          post :create, {user_id:user.to_param, :channel => { "name" => "invalid value" }}
+          post :create, {user_id:user.id, :channel => { "name" => "invalid value" }}
           expect(response).to render_template("new")
         end
       end
@@ -185,20 +185,20 @@ describe ChannelsController do
           # receives the :update_attributes message with whatever params are
           # submitted in the request.
           expect_any_instance_of(Channel).to receive(:update_attributes).with({ "name" => "MyString" })
-          put :update, {:id => channel.to_param, :channel => { "name" => "MyString" }}
+          put :update, {:id => channel.id, :channel => { "name" => "MyString" }}
         end
 
         it "assigns the requested channel as @channel" do
           channel = user.channels.create! valid_attributes
           channel = Channel.find(channel.id)
-          put :update, {:id => channel.to_param, :channel => valid_attributes}
+          put :update, {:id => channel.id, :channel => valid_attributes}
           expect(assigns(:channel)).to eq(channel)
         end
 
         it "redirects to the channel" do
           channel = user.channels.create! valid_attributes
           channel = Channel.find(channel.id)
-          put :update, {:id => channel.to_param, :channel => valid_attributes}
+          put :update, {:id => channel.id, :channel => valid_attributes}
           expect(response).to redirect_to(channel)
         end
       end
@@ -209,7 +209,7 @@ describe ChannelsController do
           channel = Channel.find(channel.id)
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Channel).to receive(:save).and_return(false)
-          put :update, {:id => channel.to_param, :channel => { "name" => "invalid value" }}
+          put :update, {:id => channel.id, :channel => { "name" => "invalid value" }}
           expect(assigns(:channel)).to eq(channel)
         end
 
@@ -217,7 +217,7 @@ describe ChannelsController do
           channel = user.channels.create! valid_attributes
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Channel).to receive(:save).and_return(false)
-          put :update, {:id => channel.to_param, :channel => { "name" => "invalid value" }}
+          put :update, {:id => channel.id, :channel => { "name" => "invalid value" }}
           expect(response).to render_template("edit")
         end
       end
@@ -226,20 +226,20 @@ describe ChannelsController do
       it "destroys the requested channel" do
         channel = user.channels.create! valid_attributes
         expect {
-          delete :destroy, {:id => channel.to_param}
+          delete :destroy, {:id => channel.id}
         }.to change(Channel, :count).by(-1)
       end
 
       it "redirects to the channels list" do
         channel = user.channels.create! valid_attributes
-        delete :destroy, {:id => channel.to_param}
+        delete :destroy, {:id => channel.id}
         expect(response).to redirect_to(user_url(user))
       end
     end
     describe "GET list_subscribers" do
       it "assigns subscribed and unsubscribed subscribers" do
         ch = create(:channel,user:user)
-        ch = Channel.find(ch)
+        ch = Channel.find(ch.id)
         subs = (0..2).map{create(:subscriber,user:user)}
         (0..1).each{|i| ch.subscribers << subs[i]}
         subscribed_subs = [subs[0],subs[1]]
@@ -252,11 +252,11 @@ describe ChannelsController do
 
       it "works when there are no subscribers yet for a channel" do
         ch = create(:channel, user:user)
-        ch = Channel.find(ch)
+        ch = Channel.find(ch.id)
         subs = (0..2).map{create(:subscriber,user:user)}
         subscribed_subs = []
         unsubscribed_subs = [subs[0],subs[1],subs[2]]
-        get :list_subscribers, {id:ch}
+        get :list_subscribers, {id: ch.id}
         expect(assigns(:channel)).to eq(ch)
         expect(assigns(:subscribed_subscribers)).to match(subscribed_subs)
         expect(assigns(:unsubscribed_subscribers).map(&:id).sort).to match(unsubscribed_subs.map(&:id).sort)
@@ -266,7 +266,7 @@ describe ChannelsController do
     describe "POST add_subscriber" do
       it "should increase subscribed subscribers array by one" do
         ch = create(:channel,user:user)
-        ch = Channel.find(ch)
+        ch = Channel.find(ch.id)
         subs = (0..2).map{create(:subscriber,user:user)}
         (0..1).each{|i| ch.subscribers << subs[i]}
         new_sub = subs[2]
@@ -276,7 +276,7 @@ describe ChannelsController do
       end
       it "should redirect to channel subscriber list" do
         ch = create(:channel,user:user)
-        ch = Channel.find(ch)
+        ch = Channel.find(ch.id)
         subs = (0..2).map{create(:subscriber,user:user)}
         (0..1).each{|i| ch.subscribers << subs[i]}
         new_sub = subs[2]
@@ -287,7 +287,7 @@ describe ChannelsController do
     describe "POST remove_subscriber" do
       it "should decrease subscribed subscribers array by one" do
         ch = create(:channel,user:user)
-        ch = Channel.find(ch)
+        ch = Channel.find(ch.id)
         subs = (0..1).map{create(:subscriber,user:user)}
         (0..1).each{|i| ch.subscribers << subs[i]}
         sub_to_remove = subs[1]
@@ -297,7 +297,7 @@ describe ChannelsController do
       end
       it "should redirect to channel subscriber list" do
         ch = create(:channel,user:user)
-        ch = Channel.find(ch)
+        ch = Channel.find(ch.id)
         subs = (0..1).map{create(:subscriber,user:user)}
         (0..1).each{|i| ch.subscribers << subs[i]}
         sub_to_remove = subs[1]
@@ -306,6 +306,4 @@ describe ChannelsController do
       end
     end
   end
-
-
 end
