@@ -11,36 +11,36 @@ describe MessagesController do
     it "is redirected to signup form always and not allowed to alter db" do
       message = channel.messages.create! valid_attributes
 
-      get :index, {channel_id:channel}
+      get :index, { channel_id: channel.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :select_import, {channel_id:channel}
+      get :select_import, { channel_id: channel.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :import, {channel_id:channel}
+      get :import, { channel_id: channel.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :show, {channel_id:channel,:id => message.to_param}
+      get :show, { channel_id: channel.id, :id => message.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :responses, {channel_id:channel,:id => message.to_param}
+      get :responses, { channel_id: channel.id, :id => message.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :new, {channel_id:channel}
+      get :new, { channel_id: channel.id }
       expect(response).to redirect_to new_user_session_path
 
-      get :edit, {channel_id:channel,:id => message.to_param}
+      get :edit, { channel_id: channel.id, :id => message.id }
       expect(response).to redirect_to new_user_session_path
 
       expect {
-            post :create, {channel_id:channel,:message => valid_attributes}
+            post :create, { channel_id: channel.id, :message => valid_attributes }
           }.to_not change(Message, :count)
 
       expect_any_instance_of(Message).not_to receive(:update_attributes)
-      put :update, {channel_id:channel,:id => message.to_param, :message => { "title" => "MyText" }}
+      put :update, {channel_id:channel.id, :id => message.id, :message => { "title" => "MyText" }}
 
       expect {
-          delete :destroy, {channel_id:channel,:id => message.to_param}
+          delete :destroy, {channel_id:channel.id, :id => message.id}
         }.to_not change(Message, :count)
     end
   end
@@ -51,36 +51,36 @@ describe MessagesController do
       another_user = create(:user)
       sign_in another_user
 
-      get :select_import, {channel_id:channel}
+      get :select_import, { channel_id: channel.id }
       expect(response).to redirect_to root_url
 
-      post :import, {channel_id:channel}
+      post :import, { channel_id:channel.id  }
       expect(response).to redirect_to root_url
 
-      get :index, {channel_id:channel}
+      get :index, { channel_id:channel.id }
       expect(response).to redirect_to root_url
 
-      get :show, {channel_id:channel,:id => message.to_param}
+      get :show, { channel_id:channel.id,:id => message.id }
       expect(response).to redirect_to root_url
 
-      get :responses, {channel_id:channel,:id => message.to_param}
+      get :responses, {channel_id:channel.id,:id => message.id}
       expect(response).to redirect_to root_url
 
-      get :new, {channel_id:channel}
+      get :new, {channel_id:channel.id}
       expect(response).to redirect_to root_url
 
-      get :edit, {channel_id:channel,:id => message.to_param}
+      get :edit, {channel_id:channel.id,:id => message.id}
       expect(response).to redirect_to root_url
 
       expect {
-            post :create, {channel_id:channel,:message => valid_attributes}
+            post :create, {channel_id:channel.id,:message => valid_attributes}
           }.to_not change(Message, :count)
 
       expect_any_instance_of(Message).not_to receive(:update_attributes)
-      put :update, {channel_id:channel,:id => message.to_param, :message => { "title" => "MyText" }}
+      put :update, {channel_id:channel.id,:id => message.id, :message => { "title" => "MyText" }}
 
       expect {
-          delete :destroy, {channel_id:channel,:id => message.to_param}
+          delete :destroy, {channel_id:channel.id,:id => message.id}
         }.to_not change(Message, :count)
 
     end
@@ -93,24 +93,24 @@ describe MessagesController do
     describe "GET index" do
       it "assigns all messages as @messages" do
         message = channel.messages.create! valid_attributes
-        get :index, {user_id:user,channel_id:channel}
-        expect(assigns(:messages)).to eq([Message.find(message)])
+        get :index, {user_id:user.id,channel_id:channel.id}
+        expect(assigns(:messages)).to eq([Message.find(message.id)])
       end
     end
 
     describe "GET show" do
       it "assigns the requested message as @message" do
         message = channel.messages.create! valid_attributes
-        get :show, {channel_id:channel,:id => message.to_param}
-        expect(assigns(:message)).to eq(Message.find(message))
+        get :show, {channel_id:channel.id,:id => message.id}
+        expect(assigns(:message)).to eq(Message.find(message.id))
       end
     end
 
     describe "GET responses" do
       it "assigns the requested message as @message" do
         message = channel.messages.create! valid_attributes
-        get :show, {channel_id:channel,:id => message.to_param}
-        expect(assigns(:message)).to eq(Message.find(message))
+        get :show, {channel_id:channel.id,:id => message.id}
+        expect(assigns(:message)).to eq(Message.find(message.id))
       end
     end
 
@@ -124,8 +124,8 @@ describe MessagesController do
     describe "GET edit" do
       it "assigns the requested message as @message" do
         message = channel.messages.create! valid_attributes
-        get :edit, {channel_id:channel,:id => message.to_param}
-        expect(assigns(:message)).to eq(Message.find(message))
+        get :edit, {channel_id:channel.id,:id => message.id}
+        expect(assigns(:message)).to eq(Message.find(message.id))
       end
     end
 
@@ -133,18 +133,18 @@ describe MessagesController do
       describe "with valid params" do
         it "creates a new Message" do
           expect {
-            post :create, {channel_id:channel,:message => valid_attributes}
+            post :create, {channel_id:channel.id,:message => valid_attributes}
           }.to change(Message, :count).by(1)
         end
 
         it "assigns a newly created message as @message" do
-          post :create, {channel_id:channel,:message => valid_attributes}
+          post :create, {channel_id:channel.id,:message => valid_attributes}
           expect(assigns(:message)).to be_a(Message)
           expect(assigns(:message)).to be_persisted
         end
 
         it "redirects to the created message" do
-          post :create, {channel_id:channel,:message => valid_attributes}
+          post :create, {channel_id:channel.id,:message => valid_attributes}
           expect(response).to redirect_to([channel,Message.last])
         end
       end
@@ -153,14 +153,14 @@ describe MessagesController do
         it "assigns a newly created but unsaved message as @message" do
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Message).to receive(:save).and_return(false)
-          post :create, {channel_id:channel,:message => { "title" => "invalid value" }}
+          post :create, {channel_id:channel.id,:message => { "title" => "invalid value" }}
           expect(assigns(:message)).to be_a_new(Message)
         end
 
         it "re-renders the 'new' template" do
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Message).to receive(:save).and_return(false)
-          post :create, {channel_id:channel,:message => { "title" => "invalid value" }}
+          post :create, {channel_id:channel.id,:message => { "title" => "invalid value" }}
           expect(response).to render_template("new")
         end
       end
@@ -175,18 +175,18 @@ describe MessagesController do
           # receives the :update_attributes message with whatever params are
           # submitted in the request.
           expect_any_instance_of(Message).to receive(:update_attributes).with({ "title" => "MyText" })
-          put :update, {channel_id:channel,:id => message.to_param, :message => { "title" => "MyText" }}
+          put :update, {channel_id:channel.id,:id => message.id, :message => { "title" => "MyText" }}
         end
 
         it "assigns the requested message as @message" do
           message = channel.messages.create! valid_attributes
-          put :update, {channel_id:channel,:id => message.to_param, :message => valid_attributes}
-          expect(assigns(:message)).to eq(Message.find(message))
+          put :update, {channel_id:channel.id,:id => message.id, :message => valid_attributes}
+          expect(assigns(:message)).to eq(Message.find(message.id))
         end
 
         it "redirects to the message" do
           message = channel.messages.create! valid_attributes
-          put :update, {channel_id:channel,:id => message.to_param, :message => valid_attributes}
+          put :update, {channel_id:channel.id,:id => message.id, :message => valid_attributes}
           expect(response).to redirect_to([channel,message])
         end
       end
@@ -196,15 +196,15 @@ describe MessagesController do
           message = channel.messages.create! valid_attributes
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Message).to receive(:save).and_return(false)
-          put :update, {channel_id:channel,:id => message.to_param, :message => { "title" => "invalid value" }}
-          expect(assigns(:message)).to eq(Message.find(message))
+          put :update, {channel_id:channel.id,:id => message.id, :message => { "title" => "invalid value" }}
+          expect(assigns(:message)).to eq(Message.find(message.id))
         end
 
         it "re-renders the 'edit' template" do
           message = channel.messages.create! valid_attributes
           # Trigger the behavior that occurs when invalid params are submitted
           allow_any_instance_of(Message).to receive(:save).and_return(false)
-          put :update, {channel_id:channel,:id => message.to_param, :message => { "title" => "invalid value" }}
+          put :update, {channel_id:channel.id,:id => message.id, :message => { "title" => "invalid value" }}
           expect(response).to render_template("edit")
         end
       end
@@ -214,13 +214,13 @@ describe MessagesController do
       it "destroys the requested message" do
         message = channel.messages.create! valid_attributes
         expect {
-          delete :destroy, {channel_id:channel,id: message.to_param}
+          delete :destroy, {channel_id:channel.id,id: message.id}
         }.to change(Message, :count).by(-1)
       end
 
       it "redirects to the channel show" do
         message = channel.messages.create! valid_attributes
-        delete :destroy, {channel_id:channel,id: message.to_param}
+        delete :destroy, {channel_id:channel.id,id: message.id}
         expect(response).to redirect_to(channel_url(channel))
       end
     end
@@ -236,14 +236,14 @@ describe MessagesController do
         expect_any_instance_of(Message).to receive(:broadcast) do |phone_nos|
           phone_nos =~ sub_nos
         end
-        post :broadcast, {channel_id:channel, id:message.to_param}
+        post :broadcast, {channel_id:channel.id, id:message.id}
       end
     end
 
     describe "GET select_import" do
       it "assigns @channel with the channel" do
-        get :select_import, {user_id:user,channel_id:channel}
-        expect(assigns(:channel)).to eq(Channel.find(channel))
+        get :select_import, {user_id:user.id, channel_id:channel.id}
+        expect(assigns(:channel)).to eq(Channel.find(channel.id))
       end
     end
 
