@@ -13,18 +13,18 @@ describe SubscriberActivityReportSender do
     sa1 = create(:subscriber_response,
       caption:"#{ch1.tparty_keyword} #{ch1.keyword} temp")
     sa2 = create(:subscriber_response,
-      caption:"#{ch2.tparty_keyword} #{ch2.keyword} temp")   
+      caption:"#{ch2.tparty_keyword} #{ch2.keyword} temp")
     sa3 = create(:subscriber_response,
-      caption:"#{ch_group.tparty_keyword} #{ch_group.keyword} temp")          
+      caption:"#{ch_group.tparty_keyword} #{ch_group.keyword} temp")
     Timecop.freeze(2014,1,27,0,55)
     expect(SubscriberActivityReportMailer).to receive(:hourly_subscriber_activity_report).exactly(3).times do |email,targets|
       case email
       when 'abc@example.com'
-        expect(targets).to match_array([Channel.find(ch1)])
+        expect(targets).to match_array([Channel.find(ch1.id)])
       when 'def@example.com'
-        expect(targets).to match_array([Channel.find(ch1)])
+        expect(targets).to match_array([Channel.find(ch1.id)])
       when 'ghi@example.com'
-        expect(targets).to match_array([Channel.find(ch1),ChannelGroup.find(ch_group)])
+        expect(targets).to match_array([Channel.find(ch1.id),ChannelGroup.find(ch_group.id)])
       else
         fail "unexpected email value #{email}"
       end
@@ -34,13 +34,13 @@ describe SubscriberActivityReportSender do
     expect(SubscriberActivityReportMailer).to receive(:daily_subscriber_activity_report).exactly(1).times do |email,targets|
       case email
       when 'def@example.com'
-        expect(targets).to match_array([Channel.find(ch2)])
+        expect(targets).to match_array([Channel.find(ch2.id)])
       else
         fail "unexpected email value #{email}"
       end
     end
     SubscriberActivityReportSender.send_daily_report
-    Timecop.return 
+    Timecop.return
   end
 
   it "send_daily_report calls mailer's corresponding method with the email and list of channels" do
